+++ src/lib/process/constants.ts
@@ -1,0 +1,3 @@
+export abstract class Constants {
+  static readonly PROCESS_ID_LENGTH: number = 32
+}
