+++ src/lib/process/map.ts
@@ -1,11 +1,9 @@
-import { Restrictions } from '../types/restrictions.js'
-
 // Changes more human-readable format of grouped restrictions to array of single restrictions for submitting to node
 // { RestrictionName: [RestrictionValue] } -> [ { RestrictionName: RestrictionValue } ]
-export const mapRestrictions = (rawRestrictions: string): Restrictions => {
+export const mapRestrictions = (rawRestrictions: string): Restrictions.Restrictions => {
   const restrictionsObj: Object = JSON.parse(rawRestrictions)
 
-  const restrictions: Restrictions = []
+  const restrictions: Restrictions.Restrictions = []
   Object.entries(restrictionsObj).map(([restrictionName, restrictionValues]) => {
     if (restrictionValues.length === 0) {
       restrictions.push({ [restrictionName]: {} })
