--- src/lib/types/options.ts
@@ -1,11 +1,0 @@
-export type Options = {
-  API_HOST: string
-  API_PORT: number
-  USER_URI: string
-}
-
-export const defaultOptions: Options = {
-  API_HOST: 'localhost',
-  API_PORT: 9944,
-  USER_URI: '//Alice',
-}
