+++ src/lib/types/polkadot.d.ts
@@ -1,0 +1,12 @@
+declare namespace Polkadot {
+  export type Options = {
+    API_HOST: string
+    API_PORT: number
+    USER_URI: string
+  }
+
+  export type Polkadot = {
+    api: import('@polkadot/api').ApiPromise
+    keyring: import('@polkadot/api').Keyring
+  }
+}
