--- src/lib/types/substrate.ts
@@ -1,6 +1,0 @@
-import { ApiPromise, Keyring } from '@polkadot/api'
-
-export type Polkadot = {
-  api: ApiPromise
-  keyring: Keyring
-}
