+++ tests/helpers/substrateHelper.ts
@@ -1,9 +1,8 @@
 import { createNodeApi } from '../../src/lib/utils/polkadot.js'
 import { getVersion, getProcess } from '../../src/lib/process/api.js'
-import { defaultOptions } from '../../src/lib/types/options.js'
-import { Process } from '../../src/lib/process/index.js'
+import { Process, defaultOptions } from '../../src/lib/process/index.js'
 
-export const getVersionHelper = async (processId: string) => {
+export const getVersionHelper = async (processId: string): Promise<number> => {
   const polkadot = await createNodeApi(defaultOptions)
   const version = await getVersion(polkadot, processId)
   await polkadot.api.disconnect()
@@ -14,5 +13,5 @@ export const getProcessHelper = async (processId: string, version: number): Prom
   const polkadot = await createNodeApi(defaultOptions)
   const process = await getProcess(polkadot, processId, version)
   await polkadot.api.disconnect()
-  return process as Process
+  return process
 }
